+++ server/elite/regions.js
@@ -9,7 +9,7 @@ const regions = require(path.resolve(base.path, 'regions.json'));
 const prefix = "regions";
 
 function getRegionByKey(key, cb) {
-	console.log("Looking for region " + key);
+    console.log("Looking for region " + key);
     // data.read(prefix, key, cb);
     let region = regions[key];
     console.log('Found ' + JSON.stringify(region));
@@ -18,8 +18,8 @@ function getRegionByKey(key, cb) {
 module.exports.getRegionByKey = getRegionByKey;
 
 function writeRegionToRedis(region) {
-	// if (region.region) {
-	// 	data.write(prefix, region.region, region);
-	// }
+    // if (region.region) {
+    //  data.write(prefix, region.region, region);
+    // }
 }
 module.exports.writeRegionToRedis = writeRegionToRedis;
