--- test/minimal-plugins.js
@@ -1,12 +1,0 @@
-var assert = require('assert');
-var plugin = require("../plugins/minimal.js");
-
-describe('minimal-plugin', function(){
-    it('should export a commands object', function(){
-        assert(typeof(plugin.commands) == 'object');
-    });
-
-    it('should export a findCommands function', function() {
-        assert(typeof(plugin.findCommand) == 'function');
-    });
-})
